♻️ Renamed VorlesungstageSelector to VorlesungstageSelect
